Extract Building class into its own file
